Use u64 indices in TreeListModel.
This simplifies interfacing between TreeListModel and the capture
database, which uses u64 indices throughout.

The GTK ListModel API only supports 32-bit row indices. We handle
this by presenting only a view of the first 2^32 rows to GTK. The
necessary clamping only needs to be implemented in a few places.

+++ src/model/mod.rs
@@ -61,7 +61,7 @@ impl GenericModel<TrafficItem> for TrafficModel {
     {
         let tree_opt  = self.imp().tree.borrow();
         let tree = tree_opt.as_ref().unwrap();
-        tree.set_expanded(self, node, position, expanded)
+        tree.set_expanded(self, node, position as u64, expanded)
     }
 
     fn update(&self) -> Result<bool, ModelError> {
@@ -95,7 +95,7 @@ impl GenericModel<DeviceItem> for DeviceModel {
     {
         let tree_opt  = self.imp().tree.borrow();
         let tree = tree_opt.as_ref().unwrap();
-        tree.set_expanded(self, node, position, expanded)
+        tree.set_expanded(self, node, position as u64, expanded)
     }
 
     fn update(&self) -> Result<bool, ModelError> {
